Removed some signed/unsigned warnings

+++ codegen/chllcode.cpp
@@ -722,7 +722,7 @@ void CHLLCode::appendExp(std::ostringstream& str, Exp *exp, PREC curPrec, bool u
 					t->getSubExp1()->isIntConst() &&
 					t->getSubExp2()->isIntConst() &&
 					((Const*)t->getSubExp2())->getInt() == 32) {
-				int sz = ((Const*)t->getSubExp1())->getInt();
+				unsigned sz = (unsigned)((Const*)t->getSubExp1())->getInt();
 				if (sz == 8 || sz == 16) {
 					bool close = false;
 					str << "*";
@@ -1311,7 +1311,8 @@ void CHLLCode::AddAssignmentStatement(int indLevel, Assign *asgn) {
 		if (rhs->getSubExp2()->isIntConst() && 
 				(((Const*)rhs->getSubExp2())->getInt() == 1 ||
 				 (lhs->getType() && lhs->getType()->isPointer() && 
-				 lhs->getType()->asPointer()->getPointsTo()->getSize() == ((Const*)rhs->getSubExp2())->getInt() * 8)))
+				 lhs->getType()->asPointer()->getPointsTo()->getSize() ==
+					(unsigned) ((Const*)rhs->getSubExp2())->getInt() * 8)))
 			s << "++";
 		else {
 			s << " += ";
